CTEST main() Restructure
A restructure for the main function for CTEST. This function exists in the
header ctest.h. The restructure was born out of a need to more quickly/easily
recognize test failure. PASS, FAIL, and SKIP indicators now appear on the
lefthand side of the test name. A more readable summary was also added.